+++ lib/payload_translator.rb
@@ -4,20 +4,6 @@ loader = Zeitwerk::Loader.for_gem
 loader.setup
 
 module PayloadTranslator
-  class Config
-    def initialize
-      @handlers = {}
-    end
-
-    def handlers
-      @handlers
-    end
-
-    def handlers=(value)
-      @handlers.merge!(value)
-    end
-  end
-
   def self.configuration
     @@config ||= Config.new
   end
