+++ spec/fixtures/with_map_formatter/input.json
@@ -1,0 +1,5 @@
+{
+  "_id": "1",
+  "name": "Jhon Doe",
+  "login_provider": "auth0"
+}
